clean up phonon-specfici helpers from app into interactive_helpers.py, generalise non-specific helpers to plot_helpers.py and clean up
