+++ ml_peg/app/utils/build_callbacks.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import base64
-from collections.abc import Callable, Iterable, Mapping, Sequence
+from collections.abc import Callable
 import io
 import json
 import math
@@ -16,7 +16,6 @@ from dash.development.base_component import Component
 from dash.exceptions import PreventUpdate
 from dash.html import Div, Iframe
 import matplotlib.pyplot as plt
-import numpy as np
 import plotly.graph_objects as go
 
 from ml_peg.analysis.utils.decorators import (
@@ -25,253 +24,6 @@ from ml_peg.analysis.utils.decorators import (
     PERIODIC_TABLE_ROWS,
 )
 from ml_peg.app.utils.weas import generate_weas_html
-
-
-def figure_from_dict(
-    figure_dict: Mapping | None, fallback_title: str | None = None
-) -> go.Figure:
-    """
-    Convert a serialized Plotly figure dictionary into a ``go.Figure``.
-
-    Parameters
-    ----------
-    figure_dict
-        JSON-serialisable figure dictionary, typically saved during analysis.
-    fallback_title
-        Title used when ``figure_dict`` is missing.
-
-    Returns
-    -------
-    go.Figure
-        Reconstructed figure with a placeholder title when necessary.
-    """
-    if not figure_dict:
-        fig = go.Figure()
-        if fallback_title:
-            fig.update_layout(title=fallback_title)
-        return fig
-    return go.Figure(figure_dict)
-
-
-def build_violin_distribution(
-    values: Sequence[float],
-    labels: Sequence[str],
-    *,
-    title: str,
-    yaxis_title: str,
-    hovertemplate: str,
-    color: str = "#636EFA",
-) -> go.Figure:
-    """
-    Build a violin plot showing a distribution of scalar errors.
-
-    Parameters
-    ----------
-    values
-        Numeric samples to plot.
-    labels
-        Labels associated with each sample, included in hover text.
-    title
-        Figure title.
-    yaxis_title
-        Axis label describing the plotted quantity.
-    hovertemplate
-        Template applied to violin samples.
-    color
-        Violin and line colour. Default corresponds to Plotly blue.
-
-    Returns
-    -------
-    go.Figure
-        Configured violin plot (or placeholder when ``values`` is empty).
-    """
-    fig = go.Figure()
-    if not values:
-        fig.update_layout(title=title or "No data available.")
-        return fig
-    customdata = [[label] for label in labels]
-    fig.add_trace(
-        go.Violin(
-            y=values,
-            text=labels,
-            customdata=customdata,
-            points="all",
-            jitter=0.05,
-            box_visible=True,
-            meanline_visible=True,
-            fillcolor=color,
-            line_color=color,
-            opacity=0.6,
-            hovertemplate=hovertemplate,
-        )
-    )
-    fig.update_layout(title=title, yaxis_title=yaxis_title, plot_bgcolor="#ffffff")
-    return fig
-
-
-def build_confusion_heatmap(
-    matrix: Iterable[Iterable[float]],
-    *,
-    x_labels: Sequence[str],
-    y_labels: Sequence[str],
-    title: str,
-    colorscale: str = "Blues",
-    hovertemplate: str | None = None,
-) -> go.Figure:
-    """
-    Build a confusion-matrix heatmap with annotated counts.
-
-    Parameters
-    ----------
-    matrix
-        2D array-like of counts.
-    x_labels
-        Labels shown along the x-axis.
-    y_labels
-        Labels shown along the y-axis.
-    title
-        Figure title.
-    colorscale
-        Plotly colourscale applied to the heatmap.
-    hovertemplate
-        Optional template overriding the default hover text.
-
-    Returns
-    -------
-    go.Figure
-        Heatmap with value annotations.
-    """
-    data = np.asarray(matrix, dtype=float)
-    fig = go.Figure()
-    if data.size == 0:
-        fig.update_layout(title=title or "No data available.", plot_bgcolor="#ffffff")
-        return fig
-    template = (
-        hovertemplate or "Pred %{y}<br>Ref %{x}<br>Count: %{z:.0f}<extra></extra>"
-    )
-    fig.add_trace(
-        go.Heatmap(
-            z=data,
-            x=x_labels,
-            y=y_labels,
-            colorscale=colorscale,
-            showscale=True,
-            hovertemplate=template,
-        )
-    )
-    max_val = data.max(initial=0.0)
-    total = data.sum()
-    for y_idx, y_label in enumerate(y_labels):
-        for x_idx, x_label in enumerate(x_labels):
-            cell_val = data[y_idx, x_idx]
-            pct = (cell_val / total * 100) if total else 0.0
-            color = "#ffffff" if max_val and cell_val >= 0.6 * max_val else "#111111"
-            fig.add_annotation(
-                x=x_label,
-                y=y_label,
-                text=f"{cell_val:.0f} ({pct:.1f}%)",
-                showarrow=False,
-                font={"color": color, "size": 14},
-            )
-    fig.update_layout(title=title, plot_bgcolor="#ffffff")
-    return fig
-
-
-def build_classified_parity_scatter(
-    points: Sequence[Mapping[str, float]],
-    *,
-    title: str,
-    xaxis_title: str,
-    yaxis_title: str,
-    class_key: str = "class",
-    ref_key: str = "ref",
-    pred_key: str = "pred",
-    label_key: str = "label",
-    hovertemplate: str,
-    colours: Mapping[str, str] | None = None,
-) -> go.Figure:
-    """
-    Build a scatter plot grouping points by classification labels.
-
-    Parameters
-    ----------
-    points
-        Sequence of metadata dictionaries containing reference/prediction values.
-    title
-        Figure title.
-    xaxis_title
-        Label for the x-axis.
-    yaxis_title
-        Label for the y-axis.
-    class_key
-        Key storing the class/group identifier. Default ``"class"``.
-    ref_key
-        Key storing reference values. Default ``"ref"``.
-    pred_key
-        Key storing predicted values. Default ``"pred"``.
-    label_key
-        Key used for hover labels. Default ``"label"``.
-    hovertemplate
-        Plotly hover template string.
-    colours
-        Optional mapping of class label -> marker colour.
-
-    Returns
-    -------
-    go.Figure
-        Scatter parity plot grouped by class.
-    """
-    fig = go.Figure()
-    if not points:
-        fig.update_layout(title=title or "No data available.")
-        return fig
-    default_colours = {
-        "TP": "#2ca02c",
-        "TN": "#1f77b4",
-        "FP": "#ff7f0e",
-        "FN": "#d62728",
-    }
-    colour_map = colours or default_colours
-    for label, colour in colour_map.items():
-        subset = [point for point in points if point.get(class_key) == label]
-        if not subset:
-            continue
-        refs = [point[ref_key] for point in subset]
-        preds = [point[pred_key] for point in subset]
-        hover = [point.get(label_key) or point.get("id") for point in subset]
-        custom = [[point.get("id") or idx] for idx, point in enumerate(subset)]
-        fig.add_trace(
-            go.Scatter(
-                x=refs,
-                y=preds,
-                mode="markers",
-                name=label,
-                marker={"color": colour},
-                text=hover,
-                customdata=custom,
-                hovertemplate=hovertemplate,
-            )
-        )
-    refs_all = [point[ref_key] for point in points]
-    preds_all = [point[pred_key] for point in points]
-    lower = min(min(refs_all), min(preds_all))
-    upper = max(max(refs_all), max(preds_all))
-    fig.add_trace(
-        go.Scatter(
-            x=[lower, upper],
-            y=[lower, upper],
-            mode="lines",
-            showlegend=False,
-            line={"color": "#8c8c8c", "dash": "dash"},
-        )
-    )
-    fig.update_layout(
-        title=title,
-        xaxis_title=xaxis_title,
-        yaxis_title=yaxis_title,
-        plot_bgcolor="#ffffff",
-    )
-    return fig
 
 
 def plot_from_table_column(
